+++ test/test_arhp_caching.rb
@@ -63,7 +63,13 @@ class ActiveRecordHostCachingTest < Minitest::Test
         end
 
         cached_db = Pool1DbC.connection.unproxied.pool.connections.first.instance_variable_get(:@_cached_current_database)
-        assert_equal("Mysql2::Error: Table '#{cached_db}.pool1_db_cs' doesn't exist", exception.message)
+
+        case ActiveRecordHostPool.loaded_db_adapter
+        when :mysql2
+          assert_equal("Mysql2::Error: Table '#{cached_db}.pool1_db_cs' doesn't exist", exception.message)
+        when :trilogy
+          assert_equal("Trilogy::ProtocolError: 1146: Table '#{cached_db}.pool1_db_cs' doesn't exist", exception.message)
+        end
       end
     end